% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_count.R
\name{test_count}
\alias{test_count}
\title{This function is used to produce the results in synthetic dataset.}
\usage{
test_count()
}
\description{
This function is used to produce the results in synthetic dataset.
}
